Start Installer construction Continue JavaDocs Create MutableDao and mutable daos implement it Add MariaDbManager to manager server process

+++ src/me/timetabler/Main.java
@@ -7,13 +7,15 @@ import javafx.stage.Stage;
 import me.timetabler.config.ConfigParser;
 import me.timetabler.config.ConfigType;
 import me.timetabler.data.dao.DaoManager;
-import me.timetabler.ui.Controller;
+import me.timetabler.ui.MainController;
+import me.timetabler.ui.main.JavaFxBridge;
 import me.util.Log;
 import me.util.LogLevel;
 import me.util.MultipleWriter;
 
 import java.io.File;
 import java.io.FileWriter;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -22,12 +24,13 @@ import java.util.Map;
 /**
  * The entry point of the program. It handles command line parameters.
  */
-public class Main extends Application{
+public class Main extends Application {
     private DaoManager daoManager;
     private static ConfigType configType = ConfigType.YAML;
 
     /**
      * Entry point to the program and handles command line parameters.
+     *
      * @param args The command line parameters.
      */
     public static void main(String[] args) {
@@ -46,7 +49,7 @@ public class Main extends Application{
     }
 
     /**
-     * Called before UI is initialised. Loads and parses all school data and initialises the map.
+     * Called before UI is initialised. Parses the config and initialises the dao manager based on the config.
      */
     @Override
     public void init() {
@@ -61,18 +64,20 @@ public class Main extends Application{
             Log.info("Initialised DaoManger");
         } catch (Exception e) {
             Log.error(e);
+            JavaFxBridge.close();
         }
     }
 
     /**
      * Initialises the UI.
+     *
      * @param primaryStage The primary JavaFX stage for the Ui to placed in.
      */
     @Override
     public void start(Stage primaryStage) {
         try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("ui/main.fxml"));
-            Controller controller = new Controller(daoManager);
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("ui/main/main.fxml"));
+            MainController controller = new MainController(daoManager);
             loader.setController(controller);
             Scene scene = new Scene(loader.load());
             primaryStage.setScene(scene);
@@ -80,16 +85,24 @@ public class Main extends Application{
             primaryStage.show();
         } catch (Exception e) {
             Log.error(e);
+            JavaFxBridge.close();
         }
     }
 
     /**
-     * Called after the UI is closed.
+     * Called after the UI is closed. It closes the daoManager.
      */
     @Override
     public void stop() {
+        daoManager.close();
     }
 
+    /**
+     * Initialises the logging mechanism, based on the given level, which can be 'v', 'd', 'i', 'w', 'e' or 'n'. Also,
+     * the log files are created and bound the logger.
+     *
+     * @param level The log level.
+     */
     private static void setupLogging(String level) {
         if ("v".equals(level)) {
             Log.LEVEL = LogLevel.VERBOSE;
@@ -109,22 +122,26 @@ public class Main extends Application{
         }
 
         try {
-            File outFile = new File("log/" + LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME).replaceAll(":", "-") + "_out.log");
-            outFile.getParentFile().mkdirs();
-            outFile.createNewFile();
+            File logFolder = new File("log/");
+            logFolder.mkdirs();
 
-            File errFile = new File("log/" + LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME).replaceAll(":", "-") + "_err.log");
-            outFile.getParentFile().mkdirs();
-            errFile.createNewFile();
+            if (!logFolder.exists() || !logFolder.canWrite()) {
+                Log.warning("Cannot create log files. Will not log to files!");
+            } else {
+                File outFile = new File(logFolder ,LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME).replaceAll(":", "-") + "_out.log");
+                File errFile = new File(logFolder, LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME).replaceAll(":", "-") + "_err.log");
+                outFile.createNewFile();
+                errFile.createNewFile();
 
-            MultipleWriter out = new MultipleWriter(new PrintWriter(System.out, true), new FileWriter(outFile, true));
-            MultipleWriter err = new MultipleWriter(new PrintWriter(System.err, true), new FileWriter(errFile, true));
+                MultipleWriter out = new MultipleWriter(new PrintWriter(System.out, true), new FileWriter(outFile, true));
+                MultipleWriter err = new MultipleWriter(new PrintWriter(System.err, true), new FileWriter(errFile, true));
 
-            Log.NORMAL_WRITER = new PrintWriter(out, true);
-            Log.ERROR_WRITER = new PrintWriter(err, true);
-        } catch (java.io.IOException e) {
-            Log.error(e);
+                Log.NORMAL_WRITER = new PrintWriter(out);
+                Log.ERROR_WRITER = new PrintWriter(err);
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+            Log.warning("Cannot create log files. Will not log to files!");
         }
     }
 }
-
